main separated to headers
Definitions, structures, and function prototypes are separated to their appropriate headers. main.c now includes these headers.
