+++ modif_ASAL/modif_ASAL.depend
@@ -11,3 +11,15 @@
 	<ctype.h>
 	<string.h>
 
+1505685646 source:e:\users\jeries\desktop\desktop files\desktop files 4-9-2017\modif_asal\modif_asal\main.c
+	<stdio.h>
+	<stdlib.h>
+	<ctype.h>
+	<string.h>
+	"double_ended_linkedlist.h"
+	"file_methods.h"
+
+1505685303 e:\users\jeries\desktop\desktop files\desktop files 4-9-2017\modif_asal\modif_asal\double_ended_linkedlist.h
+
+1505685618 e:\users\jeries\desktop\desktop files\desktop files 4-9-2017\modif_asal\modif_asal\file_methods.h
+
